next/index: move cocalc.com aspects to its separate component and hence hide GPT for on-prem
